// src/api/models/collection.ts
/* eslint-disable */
/* Code generated by ng-openapi-gen DO NOT EDIT. */

import { ApplicationUser } from '../models/application-user';
import { Customer } from '../models/customer';
import { LedgerEntry } from '../models/ledger-entry';
import { Loan } from '../models/loan';
import { PaymentMode } from '../models/payment-mode';
import { VerificationStatus } from '../models/verification-status';
export interface Collection {
  amountPaid?: number;
  collectionCode: string;
  collectionDate?: string;
  collectionId?: number;
  collector?: ApplicationUser | null;
  collectorId?: string | null;
  createdAt?: string;
  createdBy: string;
  customer?: Customer | null;
  customerId?: number;
  ledgerEntry?: LedgerEntry | null;
  loan?: Loan | null;
  loanId?: number;
  paymentMode?: PaymentMode;
  remarks?: string | null;
  updatedAt?: string;
  verificationStatus?: VerificationStatus;
}